+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Change history
 * Fixed issues with fields ``vertrouwlijkAanduiding`` and
   ``vertrouwelijkheidAanduiding``. The ZDS 1.2 specification contains invalid
   attribute names. These fields are now all named ``vertrouwelijkAanduiding``.
+* Fixed storing raw XML string when ``statustoelichting`` in service
+  ``actualiseerZaakstatus`` was empty.
 
 
 0.9.9
@@ -22,9 +24,6 @@ Change history
 
 * Fixed ``geslachtsaanduiding`` en ``geboortedatum`` to be optional for a
   ``NatuurlijkPersoon``.
-* Fixed issues with fields ``vertrouwlijkAanduiding`` and
-  ``vertrouwelijkheidAanduiding``. The ZDS 1.2 specification contains invalid
-  attribute names. These fields are now all named ``vertrouwelijkAanduiding``.
 
 
 0.9.8